Move Spring config classes to conventional package
Fix HS-1591
