+++ alert/src/main/java/org/opennms/horizon/alertservice/config/KafkaTopicProperties.java
@@ -26,30 +26,26 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.horizon.alertservice.service.routing;
+package org.opennms.horizon.alertservice.config;
 
-import org.apache.kafka.clients.admin.NewTopic;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.config.TopicBuilder;
+import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
 
-@Configuration
-public class KafkaConfig {
 
-    @Bean
-    public NewTopic alertTopic(KafkaTopicProperties kafkaTopicProperties) {
-        return getTopicBuilder(kafkaTopicProperties.getAlert()).build();
+@Getter
+@ConfigurationProperties(prefix = "kafka.topics")
+public class KafkaTopicProperties {
+    @Setter
+    private String event;
+    private final Topic alert = new Topic();
+    private final Topic monitoringPolicy = new Topic();
+
+    @Data
+    public static class Topic {
+        private String name;
+        private Integer partitions = 10;
+        private Short replicas = 1;
     }
-
-    @Bean
-    public NewTopic monitoringPolicyTopic(KafkaTopicProperties kafkaTopicProperties) {
-        return getTopicBuilder(kafkaTopicProperties.getMonitoringPolicy()).build();
-    }
-
-    private TopicBuilder getTopicBuilder(KafkaTopicProperties.Topic topic) {
-        return TopicBuilder.name(topic.getName())
-            .partitions(topic.getPartitions())
-            .replicas(topic.getReplicas());
-    }
-
 }
